refactor: Move invalid use of annotations exceptions to data.exception.annotation
