+++ src/winter/controller/FrontController.java
@@ -21,13 +21,13 @@ import winter.data.MappingMethod;
 import winter.data.client.JsonString;
 import winter.data.client.ModelView;
 import winter.data.enumdata.RequestVerb;
-import winter.data.exception.AnnotationNotFoundException;
-import winter.data.exception.DuplicateMappingException;
 import winter.data.exception.InvalidPackageNameException;
 import winter.data.exception.InvalidRequestVerbException;
 import winter.data.exception.InvalidReturnTypeException;
 import winter.data.exception.MappingNotFoundException;
 import winter.data.exception.PackageProviderNotFoundException;
+import winter.data.exception.annotation.AnnotationNotFoundException;
+import winter.data.exception.annotation.DuplicateMappingException;
 import winter.util.AnnotationScanner;
 import winter.util.ExceptionHandler;
 import winter.util.ReflectionUtil;
