+++ src/winter/data/exception/annotation/AnnotationNotFoundException.java
@@ -1,4 +1,4 @@
-package winter.data.exception;
+package winter.data.exception.annotation;
 
 public class AnnotationNotFoundException extends Exception {
     public AnnotationNotFoundException(String msg) {
