+++ src/winter/data/exception/annotation/DuplicateMappingException.java
@@ -1,4 +1,4 @@
-package winter.data.exception;
+package winter.data.exception.annotation;
 
 public class DuplicateMappingException extends RuntimeException {
     public DuplicateMappingException(String msg) {
